- Added support for using relief texture layer in advanced renderer.

+++ trunk/sources/RenderSystem/Direct3D11/spDirect3D11Query.cpp
@@ -72,7 +72,7 @@ u64 Direct3D11Query::result() const
             // Empty loop
         }
         
-        return Result;
+        return static_cast<u64>(Result);
     }
     return 0;
 }
